Refactor file handling. Support Vancity confirmation #'s.
Create new class FilterInOutFiles, which takes over the logic
of deriving an output file path from an input file path,
and of opening and closing input and output file handles
from those names. 

Instantiate FilterInOutFiles from inside OFXRepairer, so 
that its interface doesn't change.  But TODO move use of
FilterInOutFiles to the command-line app loop, and make
OFXRepairer take as parameters, open input and an output file
handles.

Refactor OFXRepairer, so that there are separate methods
to split the input file into pre, to_repair, and post; and
to perform the repair on the to_repair string. This 
structure can be re-used by future apps. 

Add a confirmation number pattern to the RE_STMTTRN, to 
accommodate present Vancity practices. 

Add doctests to the split and the repair routines. Getting
the escaping and line breaking right was hard.